requirements.txt
# requirements.txt
# Speichern Sie diesen Inhalt in einer Datei namens requirements.txt und führen Sie dann aus:
# pip install -r requirements.txt

llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.4
llama-index-llms-anthropic>=0.1.4
llama-index-llms-openai>=0.1.4
llama-index-readers-file>=0.1.4
pypdf>=3.17.0
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.2.0
sentence-transformers>=2.2.2